employeeDetails/requirements.txt
arrow==1.3.0
asgiref==3.4.1
bidict==0.23.1
blessed==1.22.0
boto3==1.18.52
botocore==1.21.52
Brotli==1.1.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2021.5.30
cffi==2.0.0
chardet==5.2.0
charset-normalizer==2.0.6
cryptography==46.0.1
cssselect2==0.8.0
Django==4.0
django-admin-list-filter-dropdown==1.0.3
django-admin-numeric-filter==0.1.6
django-admin-rangefilter==0.8.1
django-appconf==1.1.0
django-ckeditor==6.7.1
django-cors-headers==4.3.0
django-daterangefilter==1.0.0
django-easy-select2==1.5.8
django-js-asset==2.2.0
django-modeladmin-reorder==0.3.1
django-picklefield==3.2
django-q==1.3.9
django-select2==8.1.2
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docopt==0.6.2
dotenv==0.9.9
et_xmlfile==2.0.0
firebase-admin==4.3.0
fonttools==4.60.0
google-api-core==1.34.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==2.14.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.48.2
gunicorn==20.1.0
h11==0.16.0
httplib2==0.31.0
idna==3.2
jmespath==0.10.0
msgpack==1.1.1
num2words==0.5.13
numpy==1.26.4
oauth2client==4.1.3
openpyxl==3.0.10
pandas==2.2.2
pdfkit==0.6.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==3.20.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.23
pydyf==0.11.0
pyfcm==1.5.4
PyJWT==2.1.0
pyparsing==3.2.4
pyphen==0.17.2
pypng==0.0.21
PyQRCode==1.2.1
python-barcode==0.15.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2021.1
qrcode==7.4.2
redis==3.5.3
reportlab==4.3.0
requests==2.26.0
rsa==4.9.1
s3transfer==0.5.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.16.0
sqlparse==0.4.1
tinycss2==1.4.0
tinyhtml5==2.0.0
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.6
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
wsproto==1.2.0
XlsxWriter==3.1.9
zopfli==0.2.3.post1
